Show web content for each story
Allow users to navigate stories drawn from top-stories.
React Navigation is used to provide navigation functionality,
while WebView component is used to display web content of the story.

A problem here is cleartext URL.
Many of URLs provided in Hacker News are in cleartext.
App Transport Security in iOS must be disabled to show these contents.
Not sure if this would be a valid reason to disable ATS completely.
If ATS disable is not allowed, stories with cleartext URL will have
to be filtered out.